Move TunneledCheckedException to works.bosk.util
